Rename 'CustomChips' to 'CustomChip'; Remove Chips from Home screen
